+++ examples/monitor.py
@@ -21,11 +21,13 @@ from __future__ import print_function
 
 import txredisapi
 
-from twisted.application import internet
+from twisted.application.internet import ClientService, backoffPolicy
 from twisted.application import service
+from twisted.internet import reactor
+from twisted.internet.endpoints import HostnameEndpoint
 
 
-class myMonitor(txredisapi.MonitorProtocol):
+class MyMonitor(txredisapi.MonitorProtocol):
     def connectionMade(self):
         print("waiting for monitor data")
         print("use the redis client to send commands in another terminal")
@@ -38,13 +40,12 @@ class myMonitor(txredisapi.MonitorProtocol):
         print("lost connection:", reason)
 
 
-class myFactory(txredisapi.MonitorFactory):
-    # also a wapper for the ReconnectingClientFactory
-    maxDelay = 120
-    continueTrying = True
-    protocol = myMonitor
+class MyFactory(txredisapi.MonitorFactory):
+    protocol = MyMonitor
 
 
 application = service.Application("monitor")
-srv = internet.TCPClient("127.0.0.1", 6379, myFactory())
+endpoint = HostnameEndpoint(reactor, "127.0.0.1", 6379)
+srv = ClientService(endpoint, MyFactory(),
+                    retryPolicy=backoffPolicy(maxDelay=120))
 srv.setServiceParent(application)